refactor(spec_decode): keep the folded readout softmax in the model dtype
Follows the flare fusion weights, which compute torch.softmax(self.layer_fusion_weights, dim=1) directly: drop the fp32 upcast and the cast back to the activation dtype. The logits load in the model dtype, so no dtype gymnastics are needed, and an fp32 mixture output would have promoted the mixed hidden and broken the projection's dtype.

+++ vllm/model_executor/models/qwen3_domino.py
@@ -441,9 +441,9 @@ class FoldedSoftmaxReadout(nn.Module):
         chunks = hidden_states.reshape(
             *lead, self.branches, self.repeats, self.granularity
         )
-        weights = torch.softmax(self.fold_logits.float(), dim=0).to(
-            hidden_states.dtype
-        )
+        # Same dtype handling as the flare fusion weights: the logits are
+        # loaded in the model dtype, so softmax runs there directly.
+        weights = torch.softmax(self.fold_logits, dim=0)
         mixed = (
             chunks * weights.view(self.branches, 1, self.granularity)
         ).sum(dim=-3)
